Update SensorMapping.ino
GPS testing code complete along with detailed description of the 2 most common protocols. Links to wiring and descriptions on which examples and libraries to look at are in the description. I would recommend sticking to the wiring as described by the code and the websites linked. There is a TON of information that the GPS can give us and the test is currently polling it at 1Hz, which should be plenty. Everything is described in the comments of the code and this is what to expect for the rest of the sensors.