AOS, Accel, Altimeter working on Teensy
Code with detailed description on wiring on how to properly establish connections with the sensors from the Teensy using the I2C protocol. Can connect many sensors to the same pins but I show how to connect 2 to the same pins, and 1 to a different one and the syntax needed to do it.